1. update the structure of file
