--- src/main/java/muses/art/commodity/dao/CommodityCategoryDao.java
@@ -1,7 +1,0 @@
-package muses.art.commodity.dao;
-
-import muses.art.commodity.model.CommodityCategory;
-import muses.art.common.dao.BaseDao;
-
-public interface CommodityCategoryDao extends BaseDao<CommodityCategory> {
-}
