--- src/main/java/muses/art/commodity/dao/CommodityDao.java
@@ -1,7 +1,0 @@
-package muses.art.commodity.dao;
-
-import muses.art.commodity.model.Commodity;
-import muses.art.common.dao.BaseDao;
-
-public interface CommodityDao extends BaseDao<Commodity> {
-}
