--- src/main/java/muses/art/commodity/dao/ImageDao.java
@@ -1,7 +1,0 @@
-package muses.art.commodity.dao;
-
-import muses.art.commodity.model.Image;
-import muses.art.common.dao.BaseDao;
-
-public interface ImageDao extends BaseDao<Image> {
-}
