--- src/main/java/muses/art/commodity/dao/ImageSliderDao.java
@@ -1,7 +1,0 @@
-package muses.art.commodity.dao;
-
-import muses.art.commodity.model.ImageSlider;
-import muses.art.common.dao.BaseDao;
-
-public interface ImageSliderDao extends BaseDao<ImageSlider> {
-}
