+++ src/main/java/muses/art/dao/commodity/CommodityCategoryDao.java
@@ -1,0 +1,7 @@
+package muses.art.dao.commodity;
+
+import muses.art.model.commodity.CommodityCategory;
+import muses.art.dao.common.BaseDao;
+
+public interface CommodityCategoryDao extends BaseDao<CommodityCategory> {
+}
