+++ src/main/java/muses/art/dao/commodity/CommodityDao.java
@@ -1,0 +1,7 @@
+package muses.art.dao.commodity;
+
+import muses.art.model.commodity.Commodity;
+import muses.art.dao.common.BaseDao;
+
+public interface CommodityDao extends BaseDao<Commodity> {
+}
