+++ src/main/java/muses/art/dao/commodity/ImageDao.java
@@ -1,0 +1,7 @@
+package muses.art.dao.commodity;
+
+import muses.art.model.commodity.Image;
+import muses.art.dao.common.BaseDao;
+
+public interface ImageDao extends BaseDao<Image> {
+}
