+++ src/main/java/muses/art/dao/commodity/ImageSliderDao.java
@@ -1,0 +1,7 @@
+package muses.art.dao.commodity;
+
+import muses.art.model.commodity.ImageSlider;
+import muses.art.dao.common.BaseDao;
+
+public interface ImageSliderDao extends BaseDao<ImageSlider> {
+}
