+++ src/main/java/muses/art/dao/common/BaseDao.java
@@ -1,0 +1,216 @@
+package muses.art.dao.common;
+
+import java.io.Serializable;
+import java.math.BigInteger;
+import java.util.List;
+import java.util.Map;
+
+/**
+ * 基础数据库操作类
+ * <p>
+ * 其他DAO继承此类获取常用的数据库操作方法
+ *
+ * @param <T> 模型
+ */
+public interface BaseDao<T> {
+
+    /**
+     * 保存一个对象
+     *
+     * @param o 对象
+     * @return 对象的ID
+     */
+    Serializable save(T o);
+
+    /**
+     * 删除一个对象
+     *
+     * @param o 对象
+     */
+    void delete(T o);
+
+    /**
+     * 更新一个对象
+     *
+     * @param o 对象
+     */
+    void update(T o);
+
+    /**
+     * 保存或更新一个对象
+     *
+     * @param o 对象
+     */
+    void saveOrUpdate(T o);
+
+    /**
+     * 通过主键获得对象
+     *
+     * @param c  类名.class
+     * @param id 主键
+     * @return 对象
+     */
+    T get(Class<T> c, Serializable id);
+
+    /**
+     * 通过HQL语句获取一个对象
+     *
+     * @param hql HQL语句
+     * @return 对象
+     */
+    T get(String hql);
+
+    /**
+     * 通过HQL语句获取一个对象
+     *
+     * @param hql    HQL语句
+     * @param params 参数
+     * @return 对象
+     */
+    T get(String hql, Map<String, Object> params);
+
+    /**
+     * 获得对象列表
+     *
+     * @param hql HQL语句
+     * @return List
+     */
+    List<T> find(String hql);
+
+    /**
+     * 获得对象列表
+     *
+     * @param hql    HQL语句
+     * @param params 参数
+     * @return List
+     */
+    List<T> find(String hql, Map<String, Object> params);
+
+    /**
+     * 获得分页后的对象列表
+     *
+     * @param hql  HQL语句
+     * @param page 要显示第几页
+     * @param rows 每页显示多少条
+     * @return List
+     */
+    List<T> find(String hql, int page, int rows);
+
+    /**
+     * 获得分页后的对象列表
+     *
+     * @param hql    HQL语句
+     * @param params 参数
+     * @param page   要显示第几页
+     * @param rows   每页显示多少条
+     * @return List
+     */
+    List<T> find(String hql, Map<String, Object> params, int page, int rows);
+
+    /**
+     * 统计数目
+     *
+     * @param hql HQL语句(select count(*) from T)
+     * @return long
+     */
+    Long count(String hql);
+
+    /**
+     * 统计数目
+     *
+     * @param hql    HQL语句(select count(*) from T where xx = :xx)
+     * @param params 参数
+     * @return long
+     */
+    Long count(String hql, Map<String, Object> params);
+
+    /**
+     * 执行一条HQL语句
+     *
+     * @param hql HQL语句
+     * @return 响应结果数目
+     */
+    int executeHql(String hql);
+
+    /**
+     * 执行一条HQL语句
+     *
+     * @param hql    HQL语句
+     * @param params 参数
+     * @return 响应结果数目
+     */
+    int executeHql(String hql, Map<String, Object> params);
+
+    /**
+     * 获得结果集
+     *
+     * @param sql SQL语句
+     * @return 结果集
+     */
+    List<Object[]> findBySql(String sql);
+
+    /**
+     * 获得结果集
+     *
+     * @param sql  SQL语句
+     * @param page 要显示第几页
+     * @param rows 每页显示多少条
+     * @return 结果集
+     */
+    List<Object[]> findBySql(String sql, int page, int rows);
+
+    /**
+     * 获得结果集
+     *
+     * @param sql    SQL语句
+     * @param params 参数
+     * @return 结果集
+     */
+    List<Object[]> findBySql(String sql, Map<String, Object> params);
+
+    /**
+     * 获得结果集
+     *
+     * @param sql    SQL语句
+     * @param params 参数
+     * @param page   要显示第几页
+     * @param rows   每页显示多少条
+     * @return 结果集
+     */
+    List<Object[]> findBySql(String sql, Map<String, Object> params, int page, int rows);
+
+    /**
+     * 执行SQL语句
+     *
+     * @param sql SQL语句
+     * @return 响应行数
+     */
+    int executeSql(String sql);
+
+    /**
+     * 执行SQL语句
+     *
+     * @param sql    SQL语句
+     * @param params 参数
+     * @return 响应行数
+     */
+    int executeSql(String sql, Map<String, Object> params);
+
+    /**
+     * 统计
+     *
+     * @param sql SQL语句
+     * @return 数目
+     */
+    BigInteger countBySql(String sql);
+
+    /**
+     * 统计
+     *
+     * @param sql    SQL语句
+     * @param params 参数
+     * @return 数目
+     */
+    BigInteger countBySql(String sql, Map<String, Object> params);
+
+}
