+++ src/main/java/muses/art/entity/trade/Cart.java
@@ -1,0 +1,105 @@
+package muses.art.entity.trade;
+
+
+import muses.art.model.commodity.Commodity;
+import muses.art.model.user.User;
+
+import javax.persistence.*;
+import java.sql.Timestamp;
+
+@Entity
+@Table(name = "cart")
+public class Cart { // 购物车
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private int id;
+
+    @Column(name = "user_id")
+    private int userId; // 用户id(外键)
+
+    @Column(name = "commodity_id")
+    private int commodityId; // 商品id(外键)
+
+    @Column(name = "number")
+    private int number; // 数量
+
+    @Column(name = "add_num")
+    private Timestamp addTime; // 添加时间
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "user_id")
+    private User user; // 用户对象 多对一
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "commodity_id")
+    private Commodity commodity; // 商品对象 多对一
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public int getUserId() {
+        return userId;
+    }
+
+    public void setUserId(int userId) {
+        this.userId = userId;
+    }
+
+    public int getCommodityId() {
+        return commodityId;
+    }
+
+    public void setCommodityId(int commodityId) {
+        this.commodityId = commodityId;
+    }
+
+    public int getNumber() {
+        return number;
+    }
+
+    public void setNumber(int number) {
+        this.number = number;
+    }
+
+    public Timestamp getAddTime() {
+        return addTime;
+    }
+
+    public void setAddTime(Timestamp addTime) {
+        this.addTime = addTime;
+    }
+
+    public User getUser() {
+        return user;
+    }
+
+    public void setUser(User user) {
+        this.user = user;
+    }
+
+    public Commodity getCommodity() {
+        return commodity;
+    }
+
+    public void setCommodity(Commodity commodity) {
+        this.commodity = commodity;
+    }
+
+    @Override
+    public String toString() {
+        return "Cart{" +
+                "id=" + id +
+                ", userId=" + userId +
+                ", commodityId=" + commodityId +
+                ", number=" + number +
+                ", addTime=" + addTime +
+                ", user=" + user +
+                ", commodity=" + commodity +
+                '}';
+    }
+}
