+++ src/main/java/muses/art/model/commodity/Image.java
@@ -1,0 +1,64 @@
+package muses.art.model.commodity;
+
+
+import javax.persistence.*;
+
+@Entity
+@Table(name = "image")
+public class Image {
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private int id;
+
+    @Column(name = "image_url")
+    private String imageUrl; // 图像url
+
+    @Column(name = "commodity_id")
+    private int commodityId;  // 商品id(外键)
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "commodity_id")
+    private Commodity commodity; // 商品对象 多对一
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public String getImageUrl() {
+        return imageUrl;
+    }
+
+    public void setImageUrl(String imageUrl) {
+        this.imageUrl = imageUrl;
+    }
+
+    public int getCommodityId() {
+        return commodityId;
+    }
+
+    public void setCommodityId(int commodityId) {
+        this.commodityId = commodityId;
+    }
+
+    public Commodity getCommodity() {
+        return commodity;
+    }
+
+    public void setCommodity(Commodity commodity) {
+        this.commodity = commodity;
+    }
+
+    @Override
+    public String toString() {
+        return "Image{" +
+                "id=" + id +
+                ", imageUrl='" + imageUrl + '\'' +
+                ", commodityId=" + commodityId +
+                ", commodity=" + commodity +
+                '}';
+    }
+}
